+++ main.go
@@ -64,17 +64,12 @@ func main() {
 	}
 
 	k8s := kubernetes.NewForConfigOrDie(mgr.GetConfig())
-	argocd, err := services.NewArgoCDService(context.Background(), k8s , namespace, argocdRepoServer)
-	if err != nil {
-		setupLog.Error(err, "unable to connect to Argocd")
-		os.Exit(1)
-	}
 
 	if err = (&controllers.ApplicationSetReconciler{
 		Client:   mgr.GetClient(),
 		Scheme:   mgr.GetScheme(),
 		Recorder: mgr.GetEventRecorderFor("applicationset-controller"),
-		ArgocdService: argocd,
+		AppsService: services.NewArgoCDService(context.Background(), k8s , namespace, argocdRepoServer),
 	}).SetupWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create controller", "controller", "ApplicationSet")
 		os.Exit(1)
